+++ app/Helpers/Library.php
@@ -149,99 +149,6 @@ namespace App\Library;
         require_once('views/auth/view_error_message.php');
     }
 
-            // Function for declaring pagination
-    public static function paginate_declare($perpage=1) {
-      if(isset($_GET["page"])){
-          $page = intval($_GET["page"]);
-      }
-      else {
-        $page = 1;
-      }
-      $calc = $perpage * $page;
-      $start = $calc - $perpage;
-      return array("perpage"=>$perpage, "page"=>$page, "start"=>$start);
-    }    
-
-      // Funcntion for pagination
-    public static function paginate($total=1, $perpage=1, $page=1) {
-      $url = Self::getURL('&page');
-      $totalPages = ceil($total / $perpage);
-      $navigation = array('nav'=>'', 'skip_previous'=>'', 'previous'=>'', 'link'=>'', 'next'=>'', 'skip_next'=>'','endnav'=>'');
-      $navigation['nav'] = "<nav><ul class='pagination'>";
-      if($page > 1 ){
-        $navigation['skip_previous'] = "<li>
-                  <a href='$url&page=1' class='waves-effect'>
-                      |<
-                  </a>
-              </li>";
-      }
-    
-      if($page <=1 ){
-        $navigation['previous'] = " <li class='disabled'>
-                      <a href='javascript:void(0);'>
-                          Previous
-                      </a>
-                  </li>";
-      }
-      else{
-      $j = $page - 1;
-        $navigation['previous'] = " <li>
-                  <a href='$url&page=$j' class='waves-effect'>
-                      Previous
-                  </a>
-              </li>";
-      }
-    
-      if ($totalPages >=1 && $page <= $totalPages){
-        $range = 2;
-        $limit = $range+1;
-        $navigation['link'] = "";
-        if ($page > $limit){ 
-          $navigation['link'] = "<li><a href='$url&page=1' class='waves-effect'>1</a></li>
-          <li class='disabled'><a href = 'javascript:void(0);'>...</a></li>";
-        }
-            for($i= ($page-$range); $i <(($page + $range)  + 1); $i++){
-          if (($i > 0) && ($i <= $totalPages)){
-            if($i<>$page){
-              $navigation['link'] .= "<li> <a href='$url&page=$i' class='waves-effect'>$i</a></li>";
-            }
-            else{
-              $navigation['link'] .= "<li class='active'><a href='javascript:void(0);'>$i</a></li>";
-            }       
-          }
-            }
-            if ($page <= $totalPages - $limit){ 
-        $navigation['link'] .= "<li class='disabled'><a href = 'javascript:void(0);'>...</a></li>
-            <li> <a href='$url&page=" .$totalPages." ' class='waves-effect'>$totalPages</a></li>"; 
-        }
-      } 
-      if($page == $totalPages ){
-        $navigation['next'] = "<li class='disabled'>
-           <a href='javascript:void(0);'>
-                      Next
-                  </a>
-               </li>";
-      }
-      else{
-        $j = $page + 1;
-        $navigation['next'] = "<li>
-           <a href='$url&page=$j' class='waves-effect'>
-                      Next
-                  </a>
-               </li>";
-      }
-      if($page < $totalPages ){ 
-        $navigation['skip_next'] = "<li>
-                <a href='$url&page=" .$totalPages." ' class='waves-effect'>
-                    >|
-                </a>
-            </li>"; 
-      }
-      $navigation['endnav'] = "</ul></nav>";
-      return $pagination = $navigation['nav'].$navigation['skip_previous'].$navigation['previous'].$navigation['link'].$navigation['next'].$navigation['skip_next'].$navigation['endnav'];
-    }
-
-
    
 }
